refactor(admin) : table-columns 분리 , 기존 columns는 임시로 삭제하지 않은상태  #194
