+++ apps/admin/src/data/columns.tsx
@@ -17,20 +17,6 @@ import {
 } from '@appabbang/ui';
 import { createColumnHelper, type ColumnDef } from '@tanstack/react-table';
 import { SortAsc, SortDesc } from 'lucide-react';
-import {
-  formatIso,
-  formatIsoWithoutSeconds,
-  formatKR,
-  formatToDate,
-  formatToDateTime,
-} from '@/utils/format';
-import {
-  useOrderAndStatusAndDliveryTypeQuery,
-  useOrderStatusUpdateMutation,
-} from '@/hooks/use-order';
-import { usePaidUpdateMutation } from '@/hooks/use-payment';
-import { RefundDialog } from '@/components/payments/refund-dialog';
-import { useGetOrderStatusQuery } from '@/hooks/use-common-code';
 
 export interface MaterialColumns {
   no: number;
@@ -54,251 +40,6 @@ export interface PurchaseColumns {
 }
 
 export type CustomersListItem = CustomersListData[number];
-export type PaymentsListItem = PaymentsListData[number];
-export type OrderRoundListItem = OrderRoundListData[number];
-export type BreadListItem = BreadsListData[number];
-export type OrdersListItem = OrdersListData[number];
-
-export const BreadsColumns = () => {
-  const columnHelper = createColumnHelper<BreadListItem>();
-
-  const columns: ColumnDef<BreadListItem, any>[] = [
-    columnHelper.display({
-      id: 'select',
-      maxSize: 1,
-      header: ({ table }) => (
-        <Checkbox
-          className="h-5 w-5"
-          checked={
-            table.getIsAllPageRowsSelected() ||
-            (table.getIsSomePageRowsSelected() && 'indeterminate')
-          }
-          onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-        />
-      ),
-      cell: ({ row }) => (
-        <div onClick={(e) => e.preventDefault()}>
-          <Checkbox
-            className="w-5 h-5"
-            checked={row.getIsSelected()}
-            onCheckedChange={(value) => row.toggleSelected(!!value)}
-          />
-        </div>
-      ),
-    }),
-
-    columnHelper.accessor('no', {
-      maxSize: 1,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary gap-0`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 번호
-          </Button>
-        );
-      },
-      cell: (info) => {
-        const index = info.table.getPrePaginationRowModel().rows.length - info.row.index;
-        return <p className="text-center">{index}</p>;
-      },
-    }),
-
-    columnHelper.accessor('images', {
-      maxSize: 3,
-      header: '대표이미지',
-      cell: ({ row }) => {
-        const url = (row.getValue('images') as { url: string }[]) || [];
-        const src = url[0]?.url
-          ? url[0].url
-          : 'https://cdn.imweb.me/upload/S202206178ecd8851ac794/cd0f057a7035b.jpg';
-
-        return (
-          <AspectRatio ratio={9 / 5}>
-            <img
-              src={src}
-              alt={row.original.name}
-              className="h-full w-full rounded-lg object-fill"
-            />
-          </AspectRatio>
-        );
-      },
-    }),
-
-    columnHelper.accessor('name', {
-      maxSize: 3,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 메뉴명
-          </Button>
-        );
-      },
-      cell: (info) => (
-        <p className="line-clamp-3 whitespace-normal break-words">{info.getValue()}</p>
-      ),
-    }),
-
-    columnHelper.accessor('unitPrice', {
-      maxSize: 3,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 단가
-          </Button>
-        );
-      },
-      cell: (info) => (
-        <p className="text-right">
-          {new Intl.NumberFormat('ko-KR', { currency: 'KRW' }).format(info.getValue())}원
-        </p>
-      ),
-    }),
-
-    columnHelper.accessor('countryOfOrigin', {
-      maxSize: 5,
-      header: ({ column }) => '원산지정보',
-      cell: (info) => (
-        <p className="line-clamp-3 whitespace-normal break-words">{info.getValue()}</p>
-      ),
-    }),
-
-    columnHelper.accessor('allergyInfo', {
-      maxSize: 5,
-      header: ({ column }) => '알레르기 정보',
-      cell: (info) => (
-        <p className="line-clamp-3 whitespace-normal break-words">{info.getValue() || '없음'}</p>
-      ),
-    }),
-
-    columnHelper.accessor('description', {
-      maxSize: 5,
-      header: '설명',
-      cell: (info) => (
-        <p className="line-clamp-3 whitespace-normal break-words">{info.getValue()}</p>
-      ),
-    }),
-
-    // columnHelper.accessor('breadStatus', {
-    //   header: ({ column }) => {
-    //     const rawValue = column.getFilterValue();
-    //     const value = typeof rawValue === 'string' ? rawValue : 'all';
-
-    //     return (
-    //       <Select
-    //         value={value}
-    //         onValueChange={(val) => {
-    //           column.setFilterValue(val === 'all' ? undefined : val);
-    //         }}
-    //       >
-    //         <SelectTrigger>
-    //           <SelectValue />
-    //         </SelectTrigger>
-    //         <SelectContent>
-    //           <SelectItem value="all">전체</SelectItem>
-    //           {breadStatus?.map(({ name, code }) => (
-    //             <SelectItem key={code} value={code}>
-    //               {name}
-    //             </SelectItem>
-    //           ))}
-    //         </SelectContent>
-    //       </Select>
-    //     );
-    //   },
-
-    //   cell: (info) => {
-    //     const value = info.getValue();
-    //     const found = breadStatus?.find((item) => item.code === value);
-    //     const no = info.row.original.no;
-
-    //     return (
-    //       <Select
-    //         value={value}
-    //         onValueChange={(val: '10' | '20' | '30' | '40' | '50') => {
-    //           statusUpdateMutation({ no, payload: { breadStatus: val } });
-    //         }}
-    //       >
-    //         <SelectTrigger>
-    //           <SelectValue>{found?.name}</SelectValue>
-    //         </SelectTrigger>
-    //         <SelectContent>
-    //           {breadStatus?.map(({ name, code }) => (
-    //             <SelectItem key={code} value={code}>
-    //               {name}
-    //             </SelectItem>
-    //           ))}
-    //         </SelectContent>
-    //       </Select>
-    //     );
-    //   },
-
-    //   filterFn: (row, columnId, filterValue) => {
-    //     const columnValue = breadStatus?.find((item) => item.code === row.getValue(columnId));
-    //     return columnValue?.code === filterValue;
-    //   },
-    // }),
-
-    columnHelper.accessor('createdAt', {
-      maxSize: 3,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 등록일자
-          </Button>
-        );
-      },
-      cell: (info) => (
-        <p className="line-clamp-2 whitespace-normal break-words text-center">
-          {formatToDate(new Date(info.getValue()))}
-        </p>
-      ),
-    }),
-
-    columnHelper.accessor('updatedAt', {
-      maxSize: 3,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 최근 수정일
-          </Button>
-        );
-      },
-      cell: (info) => (
-        <p className="line-clamp-2 whitespace-normal break-words text-center">
-          {formatToDateTime(new Date(info.getValue()))}
-        </p>
-      ),
-    }),
-  ];
-
-  return columns;
-};
 
 export const muterialColumns = () => {
   const columnHelper = createColumnHelper<MaterialColumns>();
@@ -351,502 +92,6 @@ export const muterialColumns = () => {
         </Button>
       ),
       cell: (info) => {},
-    }),
-  ];
-
-  return columns;
-};
-
-export const ordersColumns = () => {
-  const columnHelper = createColumnHelper<OrdersListItem>();
-  const { ordersStatus } = useOrderAndStatusAndDliveryTypeQuery();
-  const { orderStatusUpdateMutation } = useOrderStatusUpdateMutation();
-
-  const columns: ColumnDef<OrdersListItem, any>[] = [
-    columnHelper.accessor('no', {
-      maxSize: 0,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary gap-0`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 번호
-          </Button>
-        );
-      },
-      cell: (info) => {
-        const index = info.table.getPrePaginationRowModel().rows.length - info.row.index;
-        return <p className="text-center">{index}</p>;
-      },
-    }),
-    columnHelper.accessor('orderNumber', {
-      maxSize: 15,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문번호
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="line-clamp-3 whitespace-normal break-words">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('customer.name', {
-      maxSize: 5,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          이름
-        </Button>
-      ),
-      cell: (info) => {
-        return info.getValue();
-      },
-    }),
-    columnHelper.accessor('customer.mobileNumber', {
-      maxSize: 10,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          전화번호
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="line-clamp-3 whitespace-normal break-words">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('payment.isPaid', {
-      maxSize: 5,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          입금확인
-        </Button>
-      ),
-      cell: (info) => {
-        const value = info.getValue() ? '완료' : '미완료';
-        return <p className={`${info.getValue() ? '' : 'text-red-500'} font-semibold`}>{value}</p>;
-      },
-    }),
-    columnHelper.accessor('orderStatus', {
-      maxSize: 5,
-      header: ({ column }) => {
-        const rawValue = column.getFilterValue();
-        const value = typeof rawValue === 'string' ? rawValue : 'all';
-
-        return (
-          <Select
-            value={value}
-            onValueChange={(val) => {
-              column.setFilterValue(val === 'all' ? undefined : val);
-            }}
-          >
-            <SelectTrigger>
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="all">전체</SelectItem>
-              {ordersStatus?.map(({ name, code }) => (
-                <SelectItem key={code} value={code}>
-                  {name}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        );
-      },
-
-      cell: (info) => {
-        const value = info.getValue();
-        const found = ordersStatus?.find((item) => item.code === value);
-        const no = info.row.original.no;
-
-        return (
-          <Select
-            value={value}
-            onValueChange={(val: '10' | '20' | '30' | '40' | '50') => {
-              orderStatusUpdateMutation({ no, orderStatus: { orderStatus: val } });
-            }}
-          >
-            <SelectTrigger>
-              <SelectValue>{found?.name}</SelectValue>
-            </SelectTrigger>
-            <SelectContent>
-              {ordersStatus?.map(({ name, code }) => (
-                <SelectItem key={code} value={code}>
-                  {name}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        );
-      },
-
-      filterFn: (row, columnId, filterValue) => {
-        const columnValue = ordersStatus?.find((item) => item.code === row.getValue(columnId));
-        return columnValue?.code === filterValue;
-      },
-    }),
-    columnHelper.accessor('deliveryMethod.name', {
-      maxSize: 5,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          배송방법
-        </Button>
-      ),
-      cell: (info) => {
-        return info.getValue();
-      },
-    }),
-    columnHelper.accessor('address', {
-      maxSize: 20,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          배송지
-        </Button>
-      ),
-      cell: (info) => {
-        const { address, addressDetail, zipcode } = info.getValue();
-
-        return (
-          <p className="line-clamp-3 whitespace-normal break-words">
-            {address} {addressDetail} ({zipcode})
-          </p>
-        );
-      },
-    }),
-    columnHelper.accessor('totalPrice', {
-      maxSize: 10,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          결제금액
-        </Button>
-      ),
-      cell: (info) => {
-        return <p>{formatKR(info.getValue())}원</p>;
-      },
-    }),
-
-    columnHelper.accessor('createdAt', {
-      maxSize: 5,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문등록일시
-        </Button>
-      ),
-      cell: (info) => {
-        return (
-          <div className="line-clamp-3 whitespace-normal break-words">
-            {formatToDate(info.getValue())}
-          </div>
-        );
-      },
-    }),
-    columnHelper.accessor('updatedAt', {
-      maxSize: 5,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문수정일시
-        </Button>
-      ),
-      cell: (info) => {
-        return (
-          <div className="line-clamp-3 whitespace-normal break-words">
-            {formatToDateTime(info.getValue())}
-          </div>
-        );
-      },
-    }),
-  ];
-
-  return columns;
-};
-
-export const orderRoundColumns = () => {
-  const columnHelper = createColumnHelper<OrderRoundListItem>();
-
-  const columns: ColumnDef<OrderRoundListItem, any>[] = [
-    columnHelper.display({
-      id: 'cell-no',
-      maxSize: 1,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary gap-0`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 번호
-          </Button>
-        );
-      },
-      cell: (info) => {
-        const index = info.table.getPrePaginationRowModel().rows.length - info.row.index;
-        return <p className="text-center">{index}</p>;
-      },
-    }),
-    columnHelper.accessor('image', {
-      maxSize: 2,
-      header: ({ column }) => <p>주문차수 이미지</p>,
-      cell: (info) => {
-        const src = info.getValue()[0]
-          ? info.getValue()[0].url
-          : 'https://cdn.imweb.me/upload/S202206178ecd8851ac794/cd0f057a7035b.jpg';
-
-        return (
-          <AspectRatio ratio={9 / 5}>
-            <img
-              src={src}
-              alt={info.row.original.name}
-              className="h-full w-full rounded-lg object-fill"
-            />
-          </AspectRatio>
-        );
-      },
-    }),
-    columnHelper.accessor('no', {
-      maxSize: 1,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문차수
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('name', {
-      maxSize: 3,
-
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          이름
-        </Button>
-      ),
-      cell: (info) => {
-        return <p>{info.getValue()}</p>;
-      },
-    }),
-
-    columnHelper.accessor('orderRoundBreads', {
-      maxSize: 6,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          판매리스트
-        </Button>
-      ),
-      cell: (info) => {
-        const names = info
-          .getValue()
-          .map((item: any) => item.name)
-          .join(', ');
-        return <p className="line-clamp-2 whitespace-normal break-words">{names}</p>;
-      },
-    }),
-    columnHelper.accessor('startedAt', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          시작일자
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{formatIsoWithoutSeconds(info.getValue())}</p>;
-      },
-    }),
-    columnHelper.accessor('endedAt', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          종료일자
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{formatIsoWithoutSeconds(info.getValue())}</p>;
-      },
-    }),
-  ];
-
-  return columns;
-};
-
-export const paymentsColumns = () => {
-  const columnHelper = createColumnHelper<PaymentsListItem>();
-  const { paidUpdateMutation } = usePaidUpdateMutation();
-  const { data: orderStatus } = useGetOrderStatusQuery();
-  const columns: ColumnDef<PaymentsListItem, any>[] = [
-    columnHelper.display({
-      id: 'cell-no',
-      maxSize: 1,
-      header: ({ column }) => {
-        const isSorted = column.getIsSorted() === 'asc';
-        return (
-          <Button
-            className={`p-0 ${isSorted ? 'text-blue-500' : ''} hover:text-primary gap-0`}
-            variant="ghost"
-            onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-          >
-            {isSorted ? <SortAsc /> : <SortDesc />} 번호
-          </Button>
-        );
-      },
-      cell: (info) => {
-        const index = info.table.getPrePaginationRowModel().rows.length - info.row.index;
-        return <p className="text-center">{index}</p>;
-      },
-    }),
-    columnHelper.accessor('order.orderNumber', {
-      maxSize: 2,
-      header: ({ column }) => <p>주문번호</p>,
-      cell: (info) => {
-        return <p>{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('order.orderStatusName', {
-      maxSize: 1,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문상태
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('isPaid', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          결제완료여부
-        </Button>
-      ),
-      cell: (info) => {
-        const value = info.getValue() ? '완료' : '미완료';
-        const no = info.row.original.no;
-        const disabledCodes = orderStatus?.filter((item) =>
-          ['접수요청', '취소요청'].includes(item.name),
-        );
-
-        const rowOrderStatus = disabledCodes?.find(
-          (item) => item.code === info.row.original.order.orderStatus,
-        );
-
-        return (
-          <Select
-            disabled={
-              !disabledCodes?.map((item) => item.code).includes(info.row.original.order.orderStatus)
-            }
-            value={value}
-            onValueChange={(selectValue) => {
-              console.log(selectValue);
-              const isPaid = selectValue === '완료' ? true : false;
-              paidUpdateMutation({
-                no,
-                rowOrderStatus: rowOrderStatus!,
-                data: { isPaid, orderNo: no },
-              });
-            }}
-          >
-            <SelectTrigger>
-              <SelectValue>{value}</SelectValue>
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value={'완료'}>완료</SelectItem>
-              <SelectItem value={'미완료'}>미완료</SelectItem>
-            </SelectContent>
-          </Select>
-        );
-      },
-    }),
-
-    columnHelper.accessor('order.totalPrice', {
-      maxSize: 6,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          결제금액
-        </Button>
-      ),
-      cell: (info) => {
-        return <p>{formatKR(info.getValue())}원</p>;
-      },
-    }),
-    columnHelper.accessor('createdAt', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          주문일시
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{formatIso(info.getValue())}</p>;
-      },
-    }),
-    columnHelper.accessor('paidConfirmedAt', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          입금확인일시
-        </Button>
-      ),
-      cell: (info) => {
-        const value = info.getValue() === null ? '미입금' : formatIso(info.getValue());
-
-        return <p className="text-center">{value}</p>;
-      },
-    }),
-    columnHelper.accessor('bankCodeName', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          은행명
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('accountNumber', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          계좌번호
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.accessor('accountHolderName', {
-      maxSize: 3,
-      header: ({ column }) => (
-        <Button className="p-0" variant="ghost">
-          예금주명
-        </Button>
-      ),
-      cell: (info) => {
-        return <p className="text-center">{info.getValue()}</p>;
-      },
-    }),
-    columnHelper.display({
-      id: 'refundDetails',
-      maxSize: 1,
-      header: ({ column }) => {
-        return <p>취소/환불정보</p>;
-      },
-      cell: (info) => {
-        const refundCode = ['50', '51', '52'];
-
-        const refund = !!refundCode.find((code) => code === info.row.original.order.orderStatus);
-        return (
-          <>
-            {refund ? (
-              <RefundDialog no={info.row.original.no} />
-            ) : (
-              <Button disabled={true}>확인</Button>
-            )}
-          </>
-        );
-      },
     }),
   ];
